Move error response into its own file
