+++ AddressesAPI.Tests/AddressesAPI.Tests.csproj
@@ -32,6 +32,7 @@
 
   <ItemGroup>
     <Folder Include="V1\Domain" />
+    <Folder Include="V2\Domain" />
   </ItemGroup>
 
 </Project>
